Update en.yml
Finally getting around to offering some updated text for the one piece we forgot to do earlier this year. Don't think this needs to be sent around for translation into other languages until we have some other stuff at some point in the future. Just thought I should get this in before the Spanish is done.