Add ability to hide trustees from the chapter page
A request from a local chapter: "while we know it's not for everyone—perhaps not even us after this trial year—really like the idea of this money being purely from Awesome. Just a group of neighbors. Not these ten people linked to their info."

This is a setting and defaults to false (i.e. the default behavior is to show the trustee section)