auto-orient images when resizing/cropping them
Images that are made with mobile phones have orientation built-in
to the image. On desktop browsers, the images will display as-cropped
(horizontally) but on mobile they will be rotated back to portrait
since the image orientation is still in the image. We pass in the
-auto-orient flag to magickly (and thus to ImageMagick) to respect
the image-defined orientation and rotate before cropping.