Fix ImageCropper::Thumbor to work with unbounded crop strings
This is a hack, really. If we see an ImageMagick crop string
indicating that we don't want to resize for images _smaller_
than a certain size, it means we don't want to scale up. Thumbor
doesn't know how to do this, so in the absence of image metadata
in these cases we just ignore the cropping completely. :(