Finalists page displays chapter name for projects from other chapters
A project from one chapter might show up in the Finalists list for
another chapter if someone from that chapter voted for it in a
particular timeframe (either in cases where a user is in multiple
chapters or if someone was just browsing someone else's projects).
This will display the name of the chapter for those projects so
we know that they're from another project during review.