Add Vote#chapter so votes stick with their chapter even if users move around
This has traditionally led to a few different problems:

- a user is in multiple chapters so the finalists page doesn't know
  which chapter the user intended the vote to be attached to
- a user moves chapters and "carries" all of their old votes over
  to the new chapter
- a user leaves a chapter and now the finalists page doesn't show
  historical project votes (since project votes to this point have
  been tied to current chapter members)

This change adds a Vote#chapter_id which is the chapter that the
vote is tied to, at the time of voting. Even if a user leaves a
chapter, the vote is still attached to the chapter for future
reference. And if a user is in multiple chapters simultaneously
then when they vote, they have multiple buttons and they can
choose which chapter they are voting for.

A user still can not vote for the same project multiple times
in multiple chapters. This is such an edge case that it wasn't worth
trying to figure out.

After running the migration, we must run VoteMigrator.new.migrate!
as a one-time conversion to to add the chapter_id to all existing votes.
There is a logic to this migration (because of the above issues,
it is not always obvious which chapter a vote should be associated with).
You can also pass in an array of explicit mappings of
1) user -> chapter explicit mappings 2) vote -> chapter explicit mappings
3) users for which the chapter_id is just the chapter_id of the
project being voted on (see the code for the exact format of what's needed).