Remove file from uppy when removing from the file list
Uppy holds on to files so you can't upload duplicates, but this is
a problem if you add a file, remove it with the "remove" button, then
decide you want to add it back again. We now hold on to the internal
uppy file ID in a data element so we can remove it.