Remove successfully uploaded files from Uppy so they don't count against Uppy's filecount
Uppy keeps track of uploaded files, which is useful when autoproceed
is off, but when it is on, then once a file is successfully uploaded
we don't want it to be in Uppy's internal upload list because those
files also count against the total number of files allowed as per the
restrictions hash. i.e. If a project has 3 images, and you upload one,
then reset the restrictions to 1 remaining, you will be unable to
upload any more because the one you just uploaded will be the "1" that
you are restricted to.