Move javascript_importmap_tags to fix Firefox loading error
For some reason, the Typekit initialization code was causing importmaps
to fail in Firefox (but not Chrome or Safari). Moving the importmaps
code above the Typekit code fixed the problem.

For reference, the error in the Firefox console was:

Uncaught TypeError: The specifier “controllers” was a bare specifier, but was not remapped to anything. Relative module specifiers must start with “./”, “../” or “/”.