Project submission confirmation email respects moderation status
Move email sending from Project after_create into the models. This
allows for better control and is just cleaner (there might be times
when we want to create a project without actually sending an email,
for import purposes for example).

* Send email immediately if the project is not held for moderation
* Send email if a project is marked as legit, but add a delay to
  give the moderator a chance to undo their decision

This is a follow-up to https://github.com/awesomefoundation/awesomebits/pull/571